Move sarif to sub commands directory
