+++ packages/plugin-sarif/src/cli.ts
@@ -1,3 +1,3 @@
-import {PluginCommand as SarifUploadCommand} from './upload'
+import {PluginCommand as SarifUploadCommand} from './commands/upload'
 
 module.exports = [SarifUploadCommand]
